refactor BatchHandle to a single class, fix mypy lint errors
