+++ airbyte/_processors/file/__init__.py
@@ -3,13 +3,14 @@
 
 from __future__ import annotations
 
-from .base import FileWriterBase, FileWriterBatchHandle
-from .jsonl import JsonlWriter
-from .parquet import ParquetWriter
+from airbyte._batch_handles import BatchHandle
+from airbyte._processors.file.base import FileWriterBase
+from airbyte._processors.file.jsonl import JsonlWriter
+from airbyte._processors.file.parquet import ParquetWriter
 
 
 __all__ = [
-    "FileWriterBatchHandle",
+    "BatchHandle",
     "FileWriterBase",
     "JsonlWriter",
     "ParquetWriter",
